Move `GoGeneAnnotationParser` to `annotations.io.go` package
