+++ phenol-analysis/src/main/java/org/monarchinitiative/phenol/analysis/GoAssociationContainer.java
@@ -3,7 +3,7 @@ package org.monarchinitiative.phenol.analysis;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.monarchinitiative.phenol.annotations.formats.go.GoGaf22Annotation;
-import org.monarchinitiative.phenol.annotations.obo.go.GoGeneAnnotationParser;
+import org.monarchinitiative.phenol.annotations.io.go.GoGeneAnnotationParser;
 import org.monarchinitiative.phenol.ontology.algo.OntologyAlgorithm;
 import org.monarchinitiative.phenol.ontology.data.Ontology;
 import org.monarchinitiative.phenol.ontology.data.Term;
