+++ phenol-analysis/src/main/java/org/monarchinitiative/phenol/analysis/TermAssociationContainer.java
@@ -2,7 +2,7 @@ package org.monarchinitiative.phenol.analysis;
 
 
 import com.google.common.collect.ImmutableList;
-import org.monarchinitiative.phenol.annotations.obo.go.GoGeneAnnotationParser;
+import org.monarchinitiative.phenol.annotations.io.go.GoGeneAnnotationParser;
 import org.monarchinitiative.phenol.base.PhenolException;
 import org.monarchinitiative.phenol.ontology.algo.OntologyAlgorithm;
 import org.monarchinitiative.phenol.ontology.data.Ontology;
