+++ phenol-annotations/src/main/java/org/monarchinitiative/phenol/annotations/io/go/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Utilities for Gene Ontology (GO) file I/O.
+ */
+package org.monarchinitiative.phenol.annotations.io.go;
