--- phenol-annotations/src/main/java/org/monarchinitiative/phenol/annotations/obo/go/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Utilities for Gene Ontology (GO) file I/O.
- */
-package org.monarchinitiative.phenol.annotations.obo.go;
